Fix dictionary issues and remove pycache files
Two issues existed with the dictionaries:
1. They did not prefix their names with the system name. This is a known
issue with Plover where it can accidentally read the main.json from the
English steno system if another system has a default dictionary called
main.json as well for example. By prefixing the names we're less likely
to hit this.

2. In the released version of the plugin, the dictionaries were not
being found. My best guess is that this is due to no MANIFEST file
existing with include_package_data set to True in the setup so
the files for the dictionary were not being copied into the build.
This is how I have solved the issue for other system plugins without
issue.

This also deletes pycache files that were included somehow.